#103 Cleanup: Refactor Game methods into Round and Phase classes
