+++ src/main/java/roborally/game/Round.java
@@ -1,0 +1,60 @@
+package roborally.game;
+
+import roborally.game.objects.gameboard.IFlag;
+import roborally.game.objects.robot.Robot;
+import roborally.ui.gdx.events.Events;
+
+import java.util.ArrayList;
+
+public class Round implements IRound {
+
+    private ArrayList<Robot> robots;
+    private IPhase phase;
+
+    public Round(Events events, ArrayList<Robot> robots, ArrayList<IFlag> flags){
+        this.robots = robots;
+        restoreRebootedRobots();
+        this.phase = new Phase(this.robots, flags, events);
+    }
+
+    @Override
+    public void announcePowerDown() {
+
+    }
+
+    @Override
+    public void dealCards() {
+
+    }
+
+    @Override
+    public void programRobots() {
+
+    }
+
+    @Override
+    public void startPhases() {
+
+    }
+
+    @Override
+    public void cleanUp() {
+
+    }
+
+    @Override
+    public void checkForDestroyedRobots() {
+
+    }
+
+    @Override
+    public void restoreRebootedRobots() {
+        for (Robot robot : robots)
+            robot.backToCheckPoint();
+    }
+
+    @Override
+    public IPhase getPhase() {
+        return this.phase;
+    }
+}
