+++ src/main/java/roborally/utilities/controls/ControlsDebug.java
@@ -15,15 +15,11 @@ public class ControlsDebug implements IControls {
         menuControlMap.put(Input.Keys.DOWN, () -> game.getFirstRobot().move(-1));
         menuControlMap.put(Input.Keys.LEFT, () -> game.getFirstRobot().rotate(Direction.turnLeftFrom(game.getFirstRobot().getLogic().getDirection())));
         menuControlMap.put(Input.Keys.RIGHT, () -> game.getFirstRobot().rotate(Direction.turnRightFrom((game.getFirstRobot().getLogic().getDirection()))));
-        menuControlMap.put(Input.Keys.F, game::fireLaser);
-        menuControlMap.put(Input.Keys.SPACE, game::registerFlagPositions);
-        menuControlMap.put(Input.Keys.W, game::checkIfSomeoneWon);
+        menuControlMap.put(Input.Keys.F, game::manuallyFireOneLaser);
         menuControlMap.put(Input.Keys.R, game::restartGame);
         menuControlMap.put(Input.Keys.ESCAPE, game::exitGame);
-        menuControlMap.put(Input.Keys.A, game::fireLasers);
         menuControlMap.put(Input.Keys.M, game.getGameOptions()::enterMenu);
-        menuControlMap.put(Input.Keys.O, game::robotPlayNextCard);
-        menuControlMap.put(Input.Keys.T, game::testEndPhase);
+        //menuControlMap.put(Input.Keys.T, game::testEndPhase);
     }
 
     @Override
@@ -37,5 +33,13 @@ public class ControlsDebug implements IControls {
 
     private void doNothing() {
         // Ok!
+    }
+
+    @Override
+    public void addInGameControls(IGame game){
+        menuControlMap.put(Input.Keys.SPACE, game.getRound().getPhase()::registerFlagPositions);
+        menuControlMap.put(Input.Keys.W, game.getRound().getPhase()::checkForWinner);
+        menuControlMap.put(Input.Keys.A, game.getRound().getPhase()::fireLasers);
+        menuControlMap.put(Input.Keys.O, game.getRound().getPhase()::robotPlayNextCard);
     }
 }
