+++ src/main/java/roborally/utilities/controls/IControls.java
@@ -1,5 +1,10 @@
 package roborally.utilities.controls;
 
+import roborally.game.IGame;
+
 public interface IControls {
+
     Runnable getAction(int keycode);
+
+    void addInGameControls(IGame game);
 }
