+++ src/main/java/roborally/utilities/enums/PhaseStep.java
@@ -7,7 +7,7 @@ public enum PhaseStep {
     MOVE_EXPRESS_CONVEYOR_BELTS,
     MOVE_COGS,
     FIRE_LASERS,
-    MOVE_BACKUP_POINTS,
+    UPDATE_CHECKPOINT,
     REGISTER_FLAG_POSITION,
     CHECK_FOR_WINNER,
     NULL_PHASE
